src/web/requirements.txt
django==2.2.10
requests==2.21.0
requests-oauthlib==1.2.0
pyjwt==1.7.1
cryptography==2.6.1
django-encrypted-model-fields==0.5.8
azure==4.0.0
azure-common==1.1.20
azure-keyvault==1.1.0
azure-mgmt==4.0.0
azure-mgmt-managementgroups==0.1.0
azure-mgmt-authorization==0.50.0
azure-mgmt-compute==4.6.2
msrestazure==0.4.34
python-dotenv==0.10.1
keyrings.alt==3.1.1
objectpath==0.6.1
azure-cli-core==2.0.70
knack==0.6.3
azure-mgmt-subscription==0.2.0
django-widget-tweaks==1.4.1
tzlocal==1.5.1
attrs==19.1.0
astroid==2.0.4
python-dateutil==2.7.3
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap3==11.1.0
